cloudback/cloudapi/requirements.txt
Django==4.2.5
django-cors-headers==4.4.0
djangorestframework==3.15.2
# mysqlclient==2.2.4
